\name{simubed}
\alias{simubed}
\title{
Simulation of bed format example file.
}
\description{
Simulation of bed format example file.
}
\usage{
simubed(nprobe=1000)
}
\arguments{
  \item{nprobe}{Number of probes on each chromosome, default is 1000}
}
\details{
Simulation of bed format example file. 
}
\value{
A data frame
}
\author{
Zongli Xu
}
\examples{
simubed(nprobe=1000)
}


